(1) integreated with the backined on Area/Zone/Sector/SectorDetails scenes. (2) The sector azimuth is working with the simulated zones/sectors location files.  (3) Sector azimuth graphics are in and working. (4) The selected sector shows sector anme and the unselected sectors show the site name.

+++ src/getDailyAverage.js
@@ -17,7 +17,7 @@
 
 function getDailyAverage(dailyAverage: string) {
   if (typeof dailyAverage === "string") {
-    return parseInt(dailyAverage);
+    return parseFloat(dailyAverage);
   }
   return dailyAverage;
 }
